Refactor extract mixins
